backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-pptx==0.6.23
pdf2image==1.17.0
openai==0.28.0
google-generativeai==0.3.2
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
sse_starlette==1.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 
